requirements.txt
flask
flask-ngrok
pillow
os